requirements.txt
streamlit
langchain
langchain-google-genai
supabase
python-dotenv
langsmith
pydantic
google-genai
